+++ packages/react/Box/src/box.types.ts
@@ -1,9 +1,7 @@
 import { HTMLAttributes, ReactNode } from "react";
-import { BoxSprinkle, box } from "@nimbus-ds/styles";
+import { BoxSprinkle } from "@nimbus-ds/styles";
 
-type Extends = Parameters<typeof box.sprinkle>[0] &
-  BoxSprinkle &
-  HTMLAttributes<HTMLDivElement>;
+type Extends = BoxSprinkle & HTMLAttributes<HTMLDivElement>;
 
 export interface BoxProps extends Extends {
   /** Element to be rendered inside the Box component */