+++ packages/react/Stack/src/stack.types.ts
@@ -1,9 +1,7 @@
 import { HTMLAttributes, ReactNode } from "react";
-import { StackSprinkle, stack } from "@nimbus-ds/styles";
+import { StackSprinkle } from "@nimbus-ds/styles";
 
-type Extends = Parameters<typeof stack.sprinkle>[0] &
-  StackSprinkle &
-  HTMLAttributes<HTMLDivElement>;
+type Extends = StackSprinkle & HTMLAttributes<HTMLDivElement>;
 
 export interface StackProps extends Extends {
   /** Element to be rendered inside the Stack component */