+++ packages/styles/src/packages/sidebar/sidebar.sprinkle.types.ts
@@ -1,0 +1,14 @@
+import { paddingProperties } from "./sidebar.sprinkle.css";
+
+type PaddingProperties = keyof typeof paddingProperties;
+
+interface Conditions<T> {
+  xs?: T;
+  md?: T;
+  lg?: T;
+}
+
+export interface SidebarSprinkle {
+  maxWidth?: string | Conditions<string>;
+  padding?: PaddingProperties | Conditions<PaddingProperties>;
+}